pyproject.toml
[project]
name = "ut-grades-dashboard"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "altair==4.2.2",
    "boto3>=1.35.0",
    "numpy>=2.3.4",
    "pandas>=2.3.3",
    "python-dotenv>=1.0.0",
]
